remove outdated (and dangerous) `feof()` call
`feof()` tests the end-of-file (`EOF`) indicator for the specified stream. It returns
a non-zero value if and only if the `EOF` indicator had been set for the specified
stream.

This function tests the stream's indicator, rather than the actual stream. This means
if you use `feof()` to control a loop's input, you (usually) end up entering the loop
one more time than necessary. It relies on another function having detected EOF in the
stream for it to actually work.

In this particular case, `fgets()` would set `EOF` for the stream, but the rest of the
loop would continue to act as if it succeeded, when in reality it didn't. Only when
the loop returns back to the `feof()` check is the file stream actually considered
"exhausted". Too late at that point.

Essentially, what `feof()` does in these (misused) cases is act like a fire alarm that
detects whether someone is in danger by looking for their respective obituary in the
newspaper, rather than look for the presence of heat and/or smoke.

relevant reading:
  - https://faq.cprogramming.com/cgi-bin/smartfaq.cgi?answer=1046476070&id=1043284351
  - https://stackoverflow.com/questions/5431941/why-is-while-feof-file-always-wrong